ingestor/requirements.txt
requests
supabase
pytz
shapely
python-dotenv
